% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgLagR.R
\name{avgLagR}
\alias{avgLagR}
\title{Average Lag Correlation over Over Measurement Interval, r_|d|}
\usage{
avgLagR(x1, x2, posdef = FALSE)
}
\arguments{
\item{x1}{first matrix to be used to estimate correlations}

\item{x2}{second matrix to be used (all variables should be the same as x1, and in the same order!!)}

\item{posdef}{adjust correlations to make the matrix positive definite? ("Bock", "Higham", and default = FALSE)}
}
\description{
Estimate the average correlation between the variables within the set over a two-wave
interval (d; generally a period of time, or number of items apart).  Diagonals will be retest
correlations over that interval.

Note: Can be converted to a sort of 'reliability-adjusted matrix' by \code{cov2cor(rd)},
which will index the degree to which correlations between variables over
measurement interval (d; generally a period of time, or number of items apart)
are smaller than the retest correlations of the variables
over that interval (see Wood, Lowman, Armstrong, & Harms, 2023)

For instance: \code{phat_xy(d) = .80} indicates that the correlation between
\code{x} and \code{y} over measurement interval t is 80\% of the retest correlations of X and Y
(i.e., their correlations with themselves) over the same measurement interval.

Importantly: \code{x1} and \code{x2} should be the same set of variables, given in the same
order, but rated at different times.  For instance, \code{x1} and \code{x2} may be the first
and second administrations of the items in some personality inventory a year
apart (or: a month, a day, etc).
}
\references{
Wood, D., Lowman, G. H., Armstrong, B. F., III, & Harms, P. D. (2023).
Using retest-adjusted correlations as indicators of the semantic similarity
of items. \emph{Journal of Personality and Social Psychology, 125}(2),
437--454. \doi{10.1037/pspp0000441}
}
