% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgRep.R
\name{avgRep}
\alias{avgRep}
\title{Averages for a Set of Repeated Items}
\usage{
avgRep(set1, set2, p = "p")
}
\arguments{
\item{set1}{first set of ratings}

\item{set2}{second set of ratings}

\item{p}{person (or subject/case) identifier; should be given as first column}
}
\value{
Person's (p) average score for variable in each column
}
\description{
Provide person-averages for each column, where repeated ratings are given
on separate rows.
}
\details{
Currently omits a person who has *any* NA for a single variable - which
would correspond to the person never rating the variable in either first or second repetition
}
