% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigcor.R
\name{bigcor}
\alias{bigcor}
\title{large correlation matrix with ff}
\usage{
bigcor(x, y = NULL, fun = c("cor", "cov"), size = 2000, verbose = TRUE, ...)
}
\arguments{
\item{x}{the matrix}

\item{verbose}{report whats going on}

\item{nblocks}{required to be factor of ncols x}
}
\description{
developed by anspiess: https://www.r-bloggers.com/2013/02/bigcor-large-correlation-matrices-in-r/
}
