% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx.R
\name{cx}
\alias{cx}
\title{Scale-Center Deviation Scores}
\usage{
cx(x, smin, smax)
}
\arguments{
\item{x}{matrix to be transformed}

\item{smin}{Scale minimum}

\item{smax}{Scale maximum}
}
\value{
Scale-centered scores on [-1,1] range.
}
\description{
Transforms scores of interest from original metric to range with minimum score
of -1 and 1.
}
