% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expRse.R
\name{expRse}
\alias{expRse}
\title{Standard Errors of Retest- or Cross-Rater-Adjusted Correlation}
\usage{
expRse(rxy, rxxyy, n)
}
\arguments{
\item{rxy}{the geometric mean of the two cross-item, cross-rater correlations (in your case, cross-item, cross-time correlations)}

\item{rxxyy}{the geometric mean of the two  cross-rater correlations of the same item (in your case, retest correlations)}

\item{n}{number of observations associated with both rxy and rxxyy}
}
\value{
standard error of reliability-adjusted rxy correlation
}
\description{
From Rene Mottus.  He writes:

'I got to this approach by applying some intuitions (like the mean of two
parallel correlations is their individual standard errors' mean divided by
sqrt(2), and then adjusted by the correlation) and then tinkering (the
initial results based on intuition alone were slightly off,
what needed tinkering was taking the roots and square roots at the right times).
It agrees well with the simulation results. If someone can tell me why then that
would be even better.'
}
