% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancyR.R
\name{fancyR}
\alias{fancyR}
\title{\href{https://github.com/fancyapps/ui}{fancyapp.js}构建的htmlwidgets小部件}
\usage{
fancyR(
  file,
  caption = NULL,
  footnote = NULL,
  data_caption = TRUE,
  thumbnail = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL,
  dots = TRUE,
  panzoom = FALSE,
  html = NULL
)
}
\arguments{
\item{file}{文件名，支持mp4,html,pdf以及其他如png等}

\item{caption}{caption}

\item{footnote}{脚注}

\item{data_caption}{每个fancybox的caption}

\item{thumbnail}{缩略图}

\item{html}{fancyapp 自定义的html布局}
}
\description{
\href{https://github.com/fancyapps/ui}{fancyapp.js}构建的htmlwidgets小部件
}
