% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertVarOrder.R
\name{invertVarOrder}
\alias{invertVarOrder}
\title{Invert a 'P-by-Order (with Vars)' Matrix to 'P-by-Var (with Orders)' Matrix}
\usage{
invertVarOrder(pByOrder)
invertVarOrder(pByOrder, vorder = pscores)
}
\arguments{
\item{pByOrder}{dataframe containing participant ID as column 1, and then columns describing
which item was rated first, second, etc}

\item{vOrder}{specify the order you would like the variables to be ordered (this is necessarily
mainly if the cells of the pByOrder matrix are variable names rather than variable NUMBERS)}
}
\value{
A participant-by-item matrix detailing when participant rated each item
}
\description{
If working with a matrix that shows which item each participant completed 1st, 2nd, etc,
it will often be more useful to invert this to show in which order the participant completed
item #1, #2, etc.
}
