% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longR.R
\name{longR}
\alias{longR}
\title{Stretch (usually a Correlation) Matrix into an Edgelist Format}
\usage{
longR(corMat)
}
\arguments{
\item{corMat}{Matrix you want to convert (doesn't need to be a correlation matrix)}
}
\value{
a dataframe with \code{c("varX","varY","rXY")} layout
}
\description{
Sometimes you just want all your correlations in one column.
}
