% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxMissing.R
\name{maxMissing}
\alias{maxMissing}
\title{Exclude Cases with N Missing Values}
\usage{
maxMissing(data, maxNA = 0, nonX)
}
\arguments{
\item{data}{Matrix to be transformed}

\item{maxNA}{Maximum number of missing values for row}

\item{nonX}{Number of leading non-score variables in datafile}
}
\value{
Reduced dataset excluding people beyond threshold missing values
}
\description{
Exclude cases that pass some specified threshold for allowable missing cases
}
\details{
The dataset must be set up to have a format of identifier variables
FOLLOWED BY (and ENDING WITH) scores in response to the inventory to be executed
correctly
}
