% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullModellavaan.R
\name{nullModellavaan}
\alias{nullModellavaan}
\title{Make a 'Null' model for MSEM analysis in lavaan}
\usage{
nullModellavaan(pmRatingData[3:ncol(pmRatingData)]) #if first two variables are 'p' and 'm' and all the remaining variables are variable ratings
}
\arguments{
\item{varset}{the set of variables you want to include in the null model}
}
\value{
the text you need to put into \code{model} for \code{lavaan::sem(model, ...)}
}
\description{
This simple little thing is just here to automate creation of 'null' models
that separate the within and between variance on scores for all variables in
a set, without then going and trying to say anything about how level1 or level2
variance is predicted by anything.
}
