% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perVar.R
\name{perVar}
\alias{perVar}
\title{Estimate Percentage Variance from Max Possible}
\usage{
perVar(data, smin, smax)
}
\arguments{
\item{data}{Matrix to be transformed (often will need to specify subset of larger dataframe)}

\item{smin}{Scale minimum}

\item{smax}{Scale maximum}
}
\value{
estimate of the proportion of the responding from max possible (range from 0 to 1)
}
\description{
For a specified set of variables (with a common range), estimate the percentage
of response variability that the person showed from maximum possible
}
\details{
This is frequently valuable to identify people who did not
vary their ratings substantially - which at extreme levels is generally indicative
of insufficient effort responding.

Note: this will currently provide inaccurate estimates for respondents
with missing values within the profile, so it is valuable to remove
respondents with large number of missing values
}
\examples{
#combine with subset function to remove people with less than 5\%
#of the maximum possible variance over this range (an indicator of invariant responding)
datafile$perVar <- perVar(datafile[varSet], 1, 5)
subdata <- subset(datafile, perVar > .05)
}
