% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phat.t.R
\name{phat.t}
\alias{phat.t}
\title{Reliability-adjusted Correlations Over a Measurement Interval (t)}
\usage{
phat.t(x1, x2)
phat.t(x1, x2, posdef = "Higham")
phat.t(x1, x2, adjust = F, posdef = "Higham") #don't adjust for retest reliability
}
\arguments{
\item{x1}{first matrix to be used to estimate correlations}

\item{x2}{second matrix to be used (all variables should be the same as x1, and in the same order!!)}

\item{adjust}{do you wish to adjust correlations for retest reliability? Defaults to TRUE}

\item{posdef}{do you wish to adjust correlations to make the matrix positive definite? ("Bock", "Higham", and defaults to FALSE)}
}
\description{
Estimate the degree to which correlations between variables over
time interval (t) are smaller than the retest correlations of the variables
over that interval.  This is estimated via an understanding of Spearman's
adjustment("correction") for measurement unreliability.

For instance: phat_xy(t) = .80 indicates that the correlation between X and Y
over measurement interval t is 80\% of the retest correlations of X and Y
(i.e., their correlations with themselves) over the same measurement interval.

Importantly: x1 and x2 should be the same set of variables, given in the same
order, but rated at different times.  For instance, x1 and x2 may be the first
and second administrations of the items in some personality inventory a year
apart (or: a month, a day, etc).
}
