% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phatd.R
\name{phatd}
\alias{phatd}
\title{Reliability-adjusted Correlations Over a Measurement Interval (d)}
\usage{
phat.t(x1, x2)
phat.t(x1, x2, posdef = "Higham")
phat.t(x1, x2, adjust = F, posdef = "Higham") #don't adjust for retest reliability
}
\arguments{
\item{x1}{first matrix to be used to estimate correlations}

\item{x2}{second matrix to be used (all variables should be the same as x1, and in the same order!!)}

\item{adjust}{adjust correlations for retest reliability? Defaults to TRUE}

\item{posdef}{adjust correlations to make the matrix positive definite? ("Bock", "Higham", and default = FALSE)}
}
\description{
Estimate the degree to which correlations between variables over
measurement interval (d; generally a period of time, or number of items apart)
are smaller than the retest correlations of the variables
over that interval.  This is estimated via an understanding of Spearman's
adjustment("correction") for measurement unreliability.

For instance: \code{phat_xy(d) = .80} indicates that the correlation between
\code{x} and \code{y} over measurement interval t is 80\% of the retest correlations of X and Y
(i.e., their correlations with themselves) over the same measurement interval.

Importantly: \code{x1} and \code{x2} should be the same set of variables, given in the same
order, but rated at different times.  For instance, \code{x1} and \code{x2} may be the first
and second administrations of the items in some personality inventory a year
apart (or: a month, a day, etc).
}
