% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomp.R
\name{pomp}
\alias{pomp}
\title{Proportion of Maximum Possible (POMP) Scores}
\usage{
pomp(x, smin, smax, p = "proportion")
}
\arguments{
\item{x}{matrix to be transformed}

\item{smin}{Scale minimum}

\item{smax}{Scale maximum}

\item{p}{Return \emph{proportion} or \emph{percent}? (default proportion)}
}
\value{
Scores with maximum possible range of [0,1] (if \strong{proportion}
specified) or [0,100] (if \strong{percent} specified)
}
\description{
Transforms scores of interest from original metric to range of 0 to 1 (if
using \strong{proportion} scaling) or range of 0 to 100 (if using
\strong{percentage} scaling.
}
