% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileAnalysis.R
\name{profileAnalysis}
\alias{profileAnalysis}
\title{Profile Correlation Analyses}
\usage{
profileAnalysis(profile1, profile2, smin, smax, qc = T)
}
\arguments{
\item{profile1}{First profile}

\item{profile2}{Second profile}

\item{smin}{Scale minimum}

\item{smax}{Scale maximum}

\item{qc}{Should scale-center correlations be estimated? (defaults to \code{T})}
}
\value{
Returns range of profile correlation results
\item{profcors}{Person-level profile correlation results}
\item{m_q_x1x2.Same}{Average profile correlation when profiles 1 and 2 are from same person}
\item{m_q_x1x2.Diff}{Average profile correlation when profiles 1 and 2 are from different people}
\item{m_q_d1d2.Same}{Average distinctive profile correlation when profiles 1 and 2 are from same person}
\item{m_q_d1d2.Diff}{Average distinctive correlation when profiles 1 and 2 are from different people}
\item{m_qAll_x1.diff}{Average profile correlation (qc) for between profile 1 between any two individuals}
\item{m_qAll_x2.diff}{Average profile correlation (qc) for between profile 2 between any two individuals}
\item{qc}{Are the results from scale-centered (i.e., \code{qc}) correlations?}
}
\description{
Estimate overall, normative, and distinctive profile correlations
between two profiles of scores (adapting procedures from Furr, 2008)
}
\details{
Note: both profiles should have the same maximum and minimum values.
}
