% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomIntModel.R
\name{randomIntModel}
\alias{randomIntModel}
\title{Make a 'Random Intercept Model' for MSEM analysis in lavaan}
\usage{
randomIntModel(pmRatingData[3:ncol(pmRatingData)]) #if first two variables
are 'p' and 'm' and all the remaining variables are variable ratings
}
\arguments{
\item{varset}{the set of variables you want to include in the null model}
}
\value{
the text you need to put into \code{model} for \code{lavaan::sem(model, ...)}.
Note that original variables \code{x} will be relabeled \code{wx} and and \code{bx} to
indicate within- and between- levels, respectively.
}
\description{
This provides code that should run a pretty agnostic 'random intercept model' in
\code{lavaan}, in much the manner detailed by Mehta & Neale (2005, Figure 3).
}
\details{
Unfortunately, there is a good chance that if you have a large number of
variables (or perhaps a large number of clusters, or observations within clusters, or...
I don't know the parameters), that this will take basically forever to run.
Unfortunately I don't have any ideas about what causes this or how to fix it.
}
