% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflectedClusters.R
\name{reflectedClusters}
\alias{reflectedClusters}
\title{Clusters with Reflected Items}
\usage{
reflectedClusters(itemRs, nclusters = 5)
}
\arguments{
\item{itemRs}{the correlation matrix including all correlations between variables in the item set}

\item{nclusters}{the number of clusters to extract. Defaults to 5.}
}
\description{
Estimate the clusters from an item set, but do so only after including
all 'reflected items'.  This allows clusters to be formed containing
antonymous content
}
