% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflectedRs.R
\name{reflectedRs}
\alias{reflectedRs}
\title{Correlation Matrix - Doubled w/ Reflections}
\usage{
reflectedRs(corrs)
}
\arguments{
\item{corrs}{the correlation matrix including all correlations between variables in the item set}
}
\description{
Artificially doubles the length of rows & columns in
correlation matrix by including all reverse-scored variable.
(Mainly used for cluster analysis)
}
