% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalebyN.R
\name{scalebyN}
\alias{scalebyN}
\title{Scale by N}
\usage{
scalebyN(x, center = T, scale = T)
}
\arguments{
\item{x}{matrix containing set of scores to be transformed}

\item{center}{mean-center scores (run \code{?scale} for more information}

\item{scale}{divide by standard deviation (run \code{?scale} for more information}
}
\value{
Standardized scores, calculated using N rather than N-1 for standard deviations
}
\description{
Estimate z-scores for column variables using N rather than standard N-1
}
\details{
\emph{Why do this?} Because it is easier to show that correlations are equivalent to (1)
 the average product of \code{zX*zY} scores (Cohen, Cohen, Aiken, & West, 2003; Eq. 2.3.1) or (2)
 the average squared difference of \code{zX} and \code{zY} scores (Cohen, Cohen, Aiken, & West, 2003; Eq. 2.2.4)
 when using z-scores computed by using \code{N} rather than \code{N-1}.  Indeed, when done so, the averages of these
 product decompositions become strictly equivalent.
}
