% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDepRDiffs.R
\name{setDepRDiffs}
\alias{setDepRDiffs}
\title{Estimate How Two (Correlated) Variables Differ in Associations with a Set of Predictors}
\usage{
setDepRDiffs(predset, dvset, controls = NULL)
}
\arguments{
\item{predset}{Set of 'Predictor' variables that will be associated with variables 2 and 3}

\item{dvset}{Set the Outcome variables will be coming from (first two columns are used as the
dependent variables to be compared)}

\item{controls}{Optional data frame of control variables. If provided, partial correlations
are computed controlling for these variables. Defaults to \code{NULL}.}
}
\value{
Most importantly: statistical significance of the differences
between all r(i,2) and r(i,3) associations
}
\description{
This basically runs the psych package's \code{r.test} function
through a set of 'predictor' variables to make it easier to find which variables
within that set differ significantly in their associations with the two variables
}
