% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiClusters.R, R/voiClusters2.R
\name{voiClusters}
\alias{voiClusters}
\title{Validity-Relevant Clusters}
\usage{
voiClusters(
  x,
  rvoi,
  threshold = 0.15,
  nk = 6,
  cname = "c",
  weight = "unit",
  controls = NULL
)

voiClusters(
  x,
  rvoi,
  threshold = 0.15,
  nk = 6,
  cname = "c",
  weight = "unit",
  controls = NULL
)
}
\arguments{
\item{x}{item set to be associated with the VOI.  It is suggested that items are placed on a [-1,1] range
pprior to analysis to decrease the variability in cluster scores produced by missing values
 (this can be done using the \code{cx} function)}

\item{rvoi}{items' correlation with the 'variable of interest' (these can be partial correlations)}

\item{threshold}{minimum correlation to be included in the 'set of relevant items'}

\item{nk}{the number of clusters you ultimately wish to extract}

\item{cname}{name of the cluster (defaults to 'c')}

\item{weight}{should items receive unit-weight in clusters (after clearing threshold),
or should they be weighted by level of association with the VOI?}
}
\value{
cluster scores, cluster membership, and full dendrogram

cluster scores, cluster membership, and full dendrogram
}
\description{
This automates a set of procedures for finding clusters of items
that relate in similar ways to a variable of interest.

This should identify clusters of items that relates somewhat independently to the variable of interest.

This automates a set of procedures for finding clusters of items
that relate in similar ways to a variable of interest.

This should produce a set of factors that relates somewhat independently to the variable of interest
}
