% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigcor.R
\name{bigcor}
\alias{bigcor}
\title{large correlation matrix with ff}
\usage{
bigcor(x, nblocks = 10, verbose = TRUE, ...)
}
\arguments{
\item{x}{the matrix}

\item{nblocks}{required to be factor of ncols x}

\item{verbose}{report whats going on}
}
\description{
developed by anspiess: https://www.r-bloggers.com/2013/02/bigcor-large-correlation-matrices-in-r/
}
