% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conScores.R
\name{conScores}
\alias{conScores}
\title{Estimated Construct Scores}
\usage{
conScores(p2i, i2c)
}
\arguments{
\item{p2i}{participant scores on the rated items (persons-to-items); It is suggested that items are on a [-1,1] range
prior to analysis to decrease the variability in cluster scores produced by missing values
 (this can be done using the \code{cx} function)}

\item{i2c}{items' coded indicativeness of the construct, on a [-1,1] metric (items-to-construct)}
}
\value{
Participant estimated construct scores, on [-1,1] range
}
\description{
Translate participant item-level scores to participant construct-level scores

Participant scores (for the p2i matrix) should be on a [-1,1] scale
prior to running this function; this can be
done through the \code{cx} function within the \code{fancyr} package
}
\examples{
# Can create the i2c variable by combining [-1,1] column indicating which items are reverse-scored with a column indicating scale the item is to be scored with:
dir<-c(1,1,1,1,-1,1,1,1,1,1,-1,1,1,-1,1,-1,-1,-1,-1,-1,1,1,-1,1)
scale<-c("scale1","scale2","scale3","scale2","scale4","scale3","scale2","scale1","scale1","scale3","scale4","scale3","scale1","scale1","scale3","scale1","scale4","scale1","scale3","scale2","scale3","scale2","scale1","scale3")
scale2<-dummy.code(scale)
i2c <- sweep(scale2, 1, dir, "*")



}
