% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conScores.R
\name{conScores}
\alias{conScores}
\title{Estimated Construct Scores}
\usage{
conScores(p2i, i2c)
}
\arguments{
\item{p2i}{participant scores on the rated items (persons-to-items); It is suggested that items are on a [-1,1] range
prior to analysis to decrease the variability in cluster scores produced by missing values
 (this can be done using the \code{cx} function)}

\item{i2c}{items' coded indicativeness of the construct, on a [-1,1] metric (items-to-construct)}
}
\value{
Participant estimated construct scores, on [-1,1] range
}
\description{
Translate participant item-level scores to participant construct-level scores

Participant scores (for the p2i matrix) should be on a [-1,1] scale
prior to running this function; this can be
done through the \code{cx} function within the \code{fancyr} package
}
