% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertVarOrder.R
\name{invertVarOrder}
\alias{invertVarOrder}
\title{Invert a 'P-Order-Var' Matrix to P-Var-Order Matrix}
\usage{
invertVarOrder(pByOrder)
}
\arguments{
\item{pByOrder}{dataframe containing participant ID as column 1, and then columns describing
which item was rated first, second, etc}
}
\value{
A participant-by-item matrix detailing when participant rated each item
}
\description{
If working with a matrix that shows which item each participant completed 1st, 2nd, etc,
it will often be more useful to invert this to show in which order the participant completed
item #1, #2, etc.
}
