% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemOrder.R
\name{itemOrder}
\alias{itemOrder}
\title{Split Item-Order Information into Separate Variables}
\usage{
itemOrder(BFIorder) #order in which person saw each BFI item, if presented randomly
}
\arguments{
\item{itemorder}{Order of which item presented the participant first, second, etc}

\item{delim}{Delimiter to use to split order string (defaults to
Qualtrics' use of '|' to separate variables)}

\item{numeric}{Convert to numeric ratings?  (Defaults to \code{T},
but may be \code{F} if itemorder presents variables names rather than numbers)}
}
\value{
Matrix of which item was presented first, second, etc)
}
\description{
Some programs, like Qualtrics, will provide item-order information in a relatively compact
single string like "##|##|##|..." to indicate order items were presented to participants

This (pretty simple) code will allow you to more easily separate this item-order information
into separate variables that can be used for quantitative analyses
}
\details{
Function indicates which items the person saw when.
This can be used to estimate item distance and such matters
}
