% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nFK.R
\name{nFK}
\alias{nFK}
\title{Number of Orthogonal Factors Measured by Set}
\usage{
nFK(rMat, rotate = "none")
}
\arguments{
\item{rMat}{Correlation matrix (with appropriate reliabilities on diagonal)
The correlation matrix should have a unique name for each variable, and these
variable names should be a single word or string (no spaces).  Can check with \code{colnames(rMat)}}
}
\value{
returns:
(1) the PCA loadings,
(2) the R^2 for each separate factor, and
(3) the nFK for the entire item set (sum of R^2)
}
\description{
For a specified set of variables, estimate the 'effective number of orthogonal factors' measured.

Note that 'measured' is intended to mean that a factor is only measured if it has an
R^2 value of 1.00.  So statistic provides the sum of R2 for orthogonal dimensions, via PCA.
}
\details{
The correlation matrix should have correct reliability values on the diagonals (ideally,
the value will be estimates of the retest values over the same measurement interval as typical for
inter-item correlations within the matrix; see Wood, Lowman, Armstrong, & Harms, 2022).
If values of 1.0 are used on the diagonal rather than correct reliability estimates,
then the resulting nFK estimate will be inflated - often substantially. However, it may
still be useful to estimate nFK in this case to provide an 'upper-bound' estimate
of nFK.

There is also some code that is a work in progress to try to do this through 'setCor', although
it is commented out as it has problems at the moment.
}
