% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nPCX.R
\name{nPCX}
\alias{nPCX}
\title{Number of Orthogonal Dimensions Measured by Set}
\usage{
nPCX(rMat)
}
\arguments{
\item{rMat}{Correlation matrix}
}
\value{
(describe)
}
\description{
For a specified set of variables, estimate the number of orthogonal dimensions measured.

Note that 'measured' is intended to mean here that a dimension is only measured if it has an
R^2 value of 1.00.  So statistic provides the sum of R2 for orthogonal dimensions, via PCA.
}
\details{
The correlation matrix should have correct reliability values on the diagonals (ideally,
the value will be estimates of the retest values over the same measurement interval as typical for
inter-item correlations within the matrix; see Wood, Lowman, Armstrong, & Harms, 2022).  If
it does not, then the resulting estimate will be inflated - often substantially!!!

There is also some code that is a work in progress to try to do this through 'setCor', although
it is commented out as it has problems at the moment.
}
\examples{
subdata <- subset(datafile, prMaxSD > .05)
}
