% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrrSplithalf.R
\name{qrrSplithalf}
\alias{qrrSplithalf}
\title{Split-Half Correlation Vector Correlation Matrix}
\usage{
qrrSplithalf(set, sims = 100, graph = FALSE, CI = 0.95, minval = -1, seed = F)
}
\arguments{
\item{set}{Set of variables and scores you wish to perform analysis on}

\item{sims}{A numeric specifying the number of random splithalves to generate.}
}
\value{
List of all separate split-half correlation vector correlation matrices
}
\description{
TBD
}
