% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflectedClusters.R, R/reflectedRs.R
\name{reflectedRs}
\alias{reflectedRs}
\title{Clusters with Reflected Items}
\usage{
reflectedRs(corrs)

reflectedRs(corrs)
}
\arguments{
\item{corrs}{Correlation matrix to reflect (or to include all reversals)}
}
\value{
Original correlation matrix 'doubled' to include all reverse-score correlations
}
\description{
Estimate the clusters from an item set, but do so only after including
all 'reflected items'.  This allows clusters to be formed containing
antonymous content

This is particularly useful for cluster analysis, where you can
include all reverse-scored variables to force the clusters to include
antonymous content (e.g., 'outgoing' and 'shy' may be on the same cluster)
}
