% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvLowVar.R
\name{rmvLowVar}
\alias{rmvLowVar}
\title{Remove Respondents Below Minimum Response Variability}
\usage{
rmvLowVar(data, smax, smin, minvar)
}
\arguments{
\item{data}{Matrix to be transformed}

\item{smax}{Scale maximum}

\item{smin}{Scale minimum}

\item{minvar}{Minimum variance allowed}
}
\value{
Dataset with cases failing to pass minimum variance screen removed
}
\description{
Exclude cases that have less than a specified minimum level of variability
}
\details{
This is a data screen designed to remove participants that did
not vary their ratings - which at extreme levels is generally indicative
of insufficient effort responding.
}
